Fixing label types.

+++ trainers/ss/ss_gan_trainer.py
@@ -23,9 +23,9 @@ class SSGanTrainer(GanTrainer):
 
     def get_ss_loss(self):
         labels_size = tf.shape(self.labels)[0]
-        real_labels = tf.concat([self.labels, tf.zeros([labels_size, 1])], axis=1)
-        fake_labels = tf.concat([tf.zeros([self.batch_size, self.classes_num], dtype=tf.float32),
-                                tf.ones([self.batch_size, 1], dtype=tf.float32)], axis=1)
+        real_labels = tf.concat([self.labels, tf.zeros([labels_size, 1], dtype=tf.uint8)], axis=1)
+        fake_labels = tf.concat([tf.zeros([self.batch_size, self.classes_num], dtype=tf.uint8),
+                                tf.ones([self.batch_size, 1], dtype=tf.uint8)], axis=1)
         # fake_labels = tf.concat([(1-alpha)*tf.ones([self.batch_size, self.classes_num])/self.classes_num,
         #                         alpha*tf.ones([self.batch_size, 1])], axis=1)
